requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-dotenv==1.4.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-tinymce==2.8.0
easy-thumbnails==2.7
googletrans==2.4.0
idna==2.8
Pillow==7.0.0
psycopg2==2.8.4
python-dotenv==0.10.5
python-memcached==1.59
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
